add strict nulls for contrib/extensions
