+++ src/sql/workbench/services/insights/browser/insightsDialogService.ts
@@ -7,7 +7,7 @@ import { createDecorator } from 'vs/platform/instantiation/common/instantiation'
 
 import { IConnectionProfile } from 'sql/platform/connection/common/interfaces';
 import { BaseActionContext } from 'sql/workbench/browser/actions';
-import { IInsightsConfigDetails, IInsightsConfig } from 'sql/platform/dashboard/browser/insightRegistry';
+import { IInsightsConfigDetails, IInsightsConfig } from 'sql/platform/extensions/common/extensions';
 
 export interface IInsightsDialogModel {
 	rows: string[][];
